[TASK] Discovery improvements
All depending classes uses the ObjectManager or uses it through Extbase DIC (fx
the SignalSlot dispatcher for the BrushRegistry slot)
